Implement integration tests for prebuilt tasks
flpath-401 https://issues.redhat.com/browse/FLPATH-401
