+++ workflow-examples/src/main/java/com/redhat/parodos/examples/JdbcWorkFlowConfiguration.java
@@ -7,10 +7,8 @@ import com.redhat.parodos.workflows.workflow.WorkFlow;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
 
 @Configuration
-@Profile("jdbc")
 public class JdbcWorkFlowConfiguration {
 
 	@Bean
