+++ workflow-examples/src/main/java/com/redhat/parodos/examples/RestWorkFlowConfiguration.java
@@ -8,10 +8,8 @@ import com.redhat.parodos.workflows.workflow.WorkFlow;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
 
 @Configuration
-@Profile("rest")
 public class RestWorkFlowConfiguration {
 
 	@Bean
